Add first pipeline/required_tasks rule
This commit adds the Dogma equivalent of HABSC Rego policy rule named
required_tasks.rego.

I'm not sure what the best option is for testing. For this commit I've
used the seedwing-policy-cli to run the verification/testing, but
perhaps it would be better to write the tests for this policy rule in
Rust instead to avoid the different input files that this one has. There
will probably be alot more of this files.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>