Add wasm32 cfg "guards"
This commit add guards around module and functions calls that use
reqwest. This is only a temporary solution to be able to make progress
on wit-bindgen.

The issue with reqwest is that when compiled with a wasm32-wasi target
it will assume that wasm-bindgen is to be used and it will generate
imports in the .wasm file produced. In our case we don't want to be
limited to a browser/web environment and instead be able to use
wit-bindgen to create a WebAssembly component that can be run on any
type of host that supports running such components.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>